Author restrictions to delete and update only own books

+++ handler/book.go
@@ -46,6 +46,7 @@ func (h *Handler) CreateBook(c *gin.Context) {
 		c.IndentedJSON(http.StatusBadRequest, gin.H{"message": "Validation error " + err.Error()})
 		return
 	}
+	book.UserId = getUserId(c)
 
 	id, err := h.service.CreateBook(book)
 	if err != nil {
@@ -74,6 +75,11 @@ func (h *Handler) UpdateBook(c *gin.Context) {
 		return
 	}
 
+	if exist.UserId != getUserId(c) {
+		c.IndentedJSON(http.StatusForbidden, gin.H{"message": "You dont have access to edit this book"})
+		return
+	}
+
 	var book entities.Book
 	if err := c.BindJSON(&book); err != nil {
 		c.IndentedJSON(http.StatusBadRequest, gin.H{"message": "Validation error " + err.Error()})
@@ -96,10 +102,36 @@ func (h *Handler) DeleteBook(c *gin.Context) {
 		return
 	}
 
+	exist, err := h.service.GetBook(id)
+	if err != nil {
+		c.IndentedJSON(http.StatusInternalServerError, gin.H{"message": "Server error " + err.Error()})
+		return
+	}
+
+	if exist.UserId != getUserId(c) {
+		c.IndentedJSON(http.StatusForbidden, gin.H{"message": "You dont have access to delete this book"})
+		return
+	}
+
 	if err := h.service.DeleteBook(id); err != nil {
 		c.IndentedJSON(http.StatusInternalServerError, gin.H{"message": "Server error " + err.Error()})
 		return
 	}
 
 	c.IndentedJSON(http.StatusOK, gin.H{"message": "Book deleted"})
+}
+
+func (h *Handler) GetUserId(c *gin.Context) {
+	id, _ := c.Get("userId")
+	c.IndentedJSON(http.StatusOK, gin.H{"userId": id})
+}
+
+func getUserId(c *gin.Context) int {
+	userId, exist := c.Get("userId")
+	if !exist {
+		c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"message": "Authorization needed"})
+		return 0
+	}
+
+	return userId.(int)
 }
